refactor : enumeration 디렉토리로 status 이동
